Implemented discount application and total price calculation in GUICalculos
