--- src/view/GUIAplicarDescuento.java
@@ -1,96 +1,0 @@
-package view;
-
-import controller.ControllerVideoJuego;
-import model.IDescuentoAplicable;
-import model.VideoJuego;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class GUIAplicarDescuento extends JFrame implements IGUIEstilos{
-
-    private JLabel idLabel, nombreLabel, titulo;
-    private JTextField idTexto, nombreTexto;
-    private JButton aplicarDescuentoBTN;
-
-    private GUIListar guiListar;
-
-    public GUIAplicarDescuento(){
-        JPanel panelFinal = new JPanel();
-        JPanel panelInvisible = new JPanel();
-        JPanel panelInvisible2 = new JPanel();
-        JPanel panelInvisible3 = new JPanel();
-        JPanel panelInvisible4 = new JPanel();
-        panelFinal.setLayout(new BorderLayout());
-        panelFinal.setBackground(COLOR);
-        panelInvisible.setPreferredSize(new Dimension(100, 150));
-        panelInvisible.setBackground(COLOR);
-        panelInvisible2.setPreferredSize(new Dimension(100, 100));
-        panelInvisible2.setBackground(COLOR);
-        panelInvisible3.setPreferredSize(new Dimension(100, 150));
-        panelInvisible3.setBackground(COLOR);
-        panelInvisible4.setPreferredSize(new Dimension(100, 150));
-        panelInvisible4.setBackground(COLOR);
-
-        titulo = new JLabel("Ingrese el Id o el Nombre del juego ");
-        Font fuenteActual = titulo.getFont();
-        titulo.setHorizontalAlignment(JLabel.CENTER);
-        titulo.setFont(new Font(fuenteActual.getName(), fuenteActual.getStyle(), 20));
-
-        JPanel panelLabels = new JPanel();
-        panelLabels.setLayout(new GridLayout(2,1));
-        panelLabels.setBackground(COLOR);
-        JPanel panelTexto = new JPanel();
-        panelTexto.setLayout(new GridLayout(2,1));
-        panelTexto.setBackground(COLOR);
-
-        setTitle("Aplicar descuento a videojuego");
-        setSize(500, 500);
-        setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-        setLayout(new BorderLayout());
-        setResizable(false);
-        setLocationRelativeTo(null);
-        getContentPane().setBackground(COLOR);
-
-        idLabel = new JLabel("Id: ");
-        idLabel.setHorizontalAlignment(JLabel.CENTER);
-        nombreLabel = new JLabel("Nombre: ");
-        nombreLabel.setHorizontalAlignment(JLabel.CENTER);
-
-        idTexto = new JTextField();
-        idTexto.setBorder(GRAY_BORDER);
-        nombreTexto = new JTextField();
-        nombreTexto.setBorder(GRAY_BORDER);
-
-        aplicarDescuentoBTN = new JButton("Aplicar Descuento");
-
-        panelLabels.add(idLabel);
-        panelLabels.add(nombreLabel);
-
-        panelTexto.add(idTexto);
-        panelTexto.add(nombreTexto);
-
-        panelFinal.add(panelInvisible, BorderLayout.NORTH);
-        panelFinal.add(panelLabels, BorderLayout.WEST);
-        panelFinal.add(panelTexto, BorderLayout.CENTER);
-        panelFinal.add(panelInvisible2, BorderLayout.EAST);
-        panelFinal.add(panelInvisible3, BorderLayout.SOUTH);
-
-        add(titulo, BorderLayout.NORTH);
-        add(panelInvisible4, BorderLayout.WEST);
-        add(panelFinal, BorderLayout.CENTER);
-        add(aplicarDescuentoBTN, BorderLayout.SOUTH);
-
-
-    }
-
-    public void setGuiListar(GUIListar guiListar) {
-        this.guiListar = guiListar;
-    }
-
-
-}
-
-
-
-
